restructure the project server-side layout
