+++ server/plugin.ts
@@ -1,3 +1,18 @@
+/*
+ * Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License").
+ * You may not use this file except in compliance with the License.
+ * A copy of the License is located at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * or in the "license" file accompanying this file. This file is distributed
+ * on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
+ * express or implied. See the License for the specific language governing
+ * permissions and limitations./routes/downloadhe License.
+ */
+
 import {
   PluginInitializerContext,
   CoreSetup,
@@ -10,7 +25,7 @@ import {
   OpendistroKibanaReportsPluginSetup,
   OpendistroKibanaReportsPluginStart,
 } from './types';
-import { defineRoutes } from './routes';
+import registerReportRoutes from './routes/report';
 
 export interface ReportsPluginRequestContext {
   logger: Logger;
@@ -39,7 +54,7 @@ export class OpendistroKibanaReportsPlugin
     const router = core.http.createRouter();
 
     // Register server side APIs
-    defineRoutes(router);
+    registerReportRoutes(router);
 
     // put logger into route handler context, so that we don't need to pass through parameters
     core.http.registerRouteHandlerContext(
